+++ requirements-dev.txt
@@ -1,0 +1,1 @@
+pre-commit==2.8.2
