+++ tests/README.md
@@ -1,0 +1,1 @@
+here will be test.py output
